pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "swbf2-image-tools"
version = "0.1.0"
description = "SWBF2 / Frosty Texture Image Tools"
readme = "README.md"
requires-python = ">=3.10,<3.14"

dependencies = [
    "pillow>=10.0",
    "pyside6>=6.6"
]

[project.scripts]
swbf2-image-tools = "SWBF2ImageTools.app:main"